+++ robot-server/tests/persistence/test_migrations.py
@@ -1,5 +1,4 @@
 """Test SQL database migrations."""
-import os
 from pathlib import Path
 from typing import Generator
 
@@ -7,7 +6,6 @@ import pytest
 import sqlalchemy
 from pytest_lazyfixture import lazy_fixture  # type: ignore[import]
 
-from robot_server.persistence import ResetManager, _CLEAR_ON_REBOOT, reset_persistence_directory
 from robot_server.persistence.database import create_sql_engine
 from robot_server.persistence.tables import (
     migration_table,
@@ -19,12 +17,6 @@ from robot_server.persistence.tables import (
 
 
 TABLES = [run_table, action_table, protocol_table, analysis_table]
-
-
-@pytest.fixture
-def reset_manager() -> ResetManager:
-    """Get a ResetManager test subject."""
-    return ResetManager()
 
 
 @pytest.fixture
@@ -88,38 +80,3 @@ def test_migration(subject: sqlalchemy.engine.Engine) -> None:
     for table in TABLES:
         values = subject.execute(sqlalchemy.select(table)).all()
         assert values == []
-
-
-async def test_test_reset_db(reset_manager: ResetManager, tmp_path: Path) -> None:
-    """Should delete persistance directory if a file makred to delete exists."""
-    assert Path(tmp_path, _CLEAR_ON_REBOOT).exists() is False
-
-    await reset_manager.reset_db(tmp_path)
-
-    assert Path(tmp_path, _CLEAR_ON_REBOOT).exists() is True
-
-
-async def test_test_reset_db_file_exist(
-    reset_manager: ResetManager, tmp_path: Path
-) -> None:
-    """Should raise an exception that the file already exists."""
-    assert Path(tmp_path, _CLEAR_ON_REBOOT).exists() is False
-
-    await reset_manager.reset_db(tmp_path)
-
-    assert Path(tmp_path, _CLEAR_ON_REBOOT).exists() is True
-
-    with pytest.raises(FileExistsError):
-        await reset_manager.reset_db(tmp_path)
-
-
-async def test_delete_persistence_directory(reset_manager: ResetManager, tmp_path: Path
-) -> None:
-    """Should make sure directory is empty."""
-    tmp_path = Path("/Users/tamarzanzouri/temp_dir/")
-
-    await reset_manager.reset_db(tmp_path)
-
-    await reset_persistence_directory(tmp_path)
-
-    assert Path(tmp_path).exists() is False
